8. Refactoring EmployeeController

+++ TestNinja/Mocking/EmployeeStorage.cs
@@ -11,8 +11,12 @@
         public void DeleteEmployee(int id)
         {
             var employee = _db.Employees.Find(id);
-            _db.Employees.Remove(employee);
-            _db.SaveChanges();
+            if (employee != null)
+            {
+                _db.Employees.Remove(employee);
+                _db.SaveChanges();
+            }
+            
         }
     }
 }
